refactor: relocate ethernet classes to spec Package modules (Rule 0007).
- ApplicationEndpoint + former NetworkEndpoint.py classes -> EthernetTopology.py
  (spec Package EthernetTopology per Tables 6.124/6.139/6.144)
- SocketConnection/SoAdRoutingGroup cluster -> new ObsoleteModel.py
  (spec Package ObsoleteModel per Tables F.115/F.116); SocketConnection is
  NOT XSD-only - Table F.116 exists in the R23-11 obsolete appendix
- CanControllerConfiguration/CanXlProps -> Fibex4Can/CanTopology.py (Table 3.14)
- delete NetworkEndpoint.py / EthernetCommunication.py (non-spec module names)
- update parser/writer/ARPackage/CoreTopology imports + test mirrors; break
  ServiceInstances<->EthernetTopology and CoreTopology<->EthernetTopology cycles
  with function-level imports (existing codebase pattern)
