+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/Fibex4Can/CanTopology.py
@@ -7,6 +7,7 @@ from typing import Optional
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import Boolean, Float, Integer, PositiveInteger
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import PositiveUnlimitedInteger, TimeValue
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
+from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.FibexCore.CoreTopology import CommunicationConnector, CommunicationController, PhysicalChannel
 
 
@@ -1055,4 +1056,173 @@ class CanCommunicationConnector(AbstractCanCommunicationConnector):
     def setPncWakeupDlc(self, value):
         if value is not None:
             self.pncWakeupDlc = value
+        return self
+
+
+class CanControllerConfiguration(ARObject):
+    """
+    CAN 2.0 configuration parameters for the CAN XL controller.
+
+    Placeholder for the AUTOSAR CAN-CONTROLLER-CONFIGURATION meta-class
+    (AUTOSAR_CP_TPS_SystemTemplate, Table 3.14). The inner attributes
+    (propSeg, syncJumpWidth, timeSeg1, timeSeg2) are not yet modeled; this
+    class is provided so CanXlProps.canConfig can be referenced. To be fully
+    synced in a later pass.
+    """
+
+    def __init__(self):
+        super().__init__()
+
+
+class CanXlProps(ARElement):
+    """
+    This meta-class is used to configure Machine specific CAN XL attributes.
+    """
+
+    # CanXlProps method parity checklist:
+    # Spec: AUTOSAR_AP_TPS_SystemDesign (AdaptivePlatform), class CAN-XL-PROPS, p.n/a (source: AUTOSAR_00052.xsd)
+    # Columns: impl / docstring / test / reader / writer   ([—] = no XML element)
+    # [x] __init__                       [x] impl  [x] docstring  [x] test  [—] reader  [—] writer
+    # [x] getCanBaudrate                 [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanBaudrate                 [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanConfig                   [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanConfig                   [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanFdBaudrate               [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanFdBaudrate               [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanFdConfig                 [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanFdConfig                 [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanXlBaudrate               [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanXlBaudrate               [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanXlConfig                 [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanXlConfig                 [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+    # [x] getCanXlConfigReqs             [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
+    # [x] setCanXlConfigReqs             [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
+
+    def __init__(self, parent: ARObject, short_name: str):
+        super().__init__(parent, short_name)
+
+        # Specifies the data segment CAN 2.0 baud rate of the CAN XL controller in bits/s.
+        self.canBaudrate: Optional[PositiveInteger] = None
+
+        # CAN 2.0 configuration parameters for the CAN XL controller.
+        self.canConfig: Optional[CanControllerConfiguration] = None
+
+        # Specifies the data segment CAN FD baud rate of the CAN XL controller in bits/s.
+        self.canFdBaudrate: Optional[PositiveInteger] = None
+
+        # CAN FD configuration parameters for the CAN XL controller.
+        self.canFdConfig: Optional[CanControllerFdConfiguration] = None
+
+        # Specifies the data segment CAN XL baud rate of the CAN XL controller in bits/s.
+        self.canXlBaudrate: Optional[PositiveInteger] = None
+
+        # CAN XL configuration parameters for the CAN XL controller.
+        self.canXlConfig: Optional[CanControllerXlConfiguration] = None
+
+        # CAN XL configuration parameter requirements for the CAN XL controller.
+        self.canXlConfigReqs: Optional[CanControllerXlConfigurationRequirements] = None
+
+    def getCanBaudrate(self) -> Optional[PositiveInteger]:
+        """
+        Specifies the data segment CAN 2.0 baud rate of the CAN XL controller in bits/s.
+        """
+        return self.canBaudrate
+
+    def setCanBaudrate(self, value: Optional[PositiveInteger]) -> "CanXlProps":
+        """
+        Specifies the data segment CAN 2.0 baud rate of the CAN XL controller in bits/s.
+        A None value is a no-op and does not overwrite an existing canBaudrate.
+        """
+        if value is not None:
+            self.canBaudrate = value
+        return self
+
+    def getCanConfig(self) -> Optional[CanControllerConfiguration]:
+        """
+        CAN 2.0 configuration parameters for the CAN XL controller.
+        """
+        return self.canConfig
+
+    def setCanConfig(self, value: Optional[CanControllerConfiguration]) -> "CanXlProps":
+        """
+        CAN 2.0 configuration parameters for the CAN XL controller.
+        A None value is a no-op and does not overwrite an existing canConfig.
+        """
+        if value is not None:
+            self.canConfig = value
+        return self
+
+    def getCanFdBaudrate(self) -> Optional[PositiveInteger]:
+        """
+        Specifies the data segment CAN FD baud rate of the CAN XL controller in bits/s.
+        """
+        return self.canFdBaudrate
+
+    def setCanFdBaudrate(self, value: Optional[PositiveInteger]) -> "CanXlProps":
+        """
+        Specifies the data segment CAN FD baud rate of the CAN XL controller in bits/s.
+        A None value is a no-op and does not overwrite an existing canFdBaudrate.
+        """
+        if value is not None:
+            self.canFdBaudrate = value
+        return self
+
+    def getCanFdConfig(self) -> Optional[CanControllerFdConfiguration]:
+        """
+        CAN FD configuration parameters for the CAN XL controller.
+        """
+        return self.canFdConfig
+
+    def setCanFdConfig(self, value: Optional[CanControllerFdConfiguration]) -> "CanXlProps":
+        """
+        CAN FD configuration parameters for the CAN XL controller.
+        A None value is a no-op and does not overwrite an existing canFdConfig.
+        """
+        if value is not None:
+            self.canFdConfig = value
+        return self
+
+    def getCanXlBaudrate(self) -> Optional[PositiveInteger]:
+        """
+        Specifies the data segment CAN XL baud rate of the CAN XL controller in bits/s.
+        """
+        return self.canXlBaudrate
+
+    def setCanXlBaudrate(self, value: Optional[PositiveInteger]) -> "CanXlProps":
+        """
+        Specifies the data segment CAN XL baud rate of the CAN XL controller in bits/s.
+        A None value is a no-op and does not overwrite an existing canXlBaudrate.
+        """
+        if value is not None:
+            self.canXlBaudrate = value
+        return self
+
+    def getCanXlConfig(self) -> Optional[CanControllerXlConfiguration]:
+        """
+        CAN XL configuration parameters for the CAN XL controller.
+        """
+        return self.canXlConfig
+
+    def setCanXlConfig(self, value: Optional[CanControllerXlConfiguration]) -> "CanXlProps":
+        """
+        CAN XL configuration parameters for the CAN XL controller.
+        A None value is a no-op and does not overwrite an existing canXlConfig.
+        """
+        if value is not None:
+            self.canXlConfig = value
+        return self
+
+    def getCanXlConfigReqs(self) -> Optional[CanControllerXlConfigurationRequirements]:
+        """
+        CAN XL configuration parameter requirements for the CAN XL controller.
+        """
+        return self.canXlConfigReqs
+
+    def setCanXlConfigReqs(self, value: Optional[CanControllerXlConfigurationRequirements]) -> "CanXlProps":
+        """
+        CAN XL configuration parameter requirements for the CAN XL controller.
+        A None value is a no-op and does not overwrite an existing canXlConfigReqs.
+        """
+        if value is not None:
+            self.canXlConfigReqs = value
         return self
