+++ src/armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/Fibex4Ethernet/ServiceInstances.py
@@ -18,8 +18,8 @@ from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.Identifiable import ARElement, Identifiable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.TagWithOptionalValue import TagWithOptionalValue
-from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetCommunication import SocketConnection, SocketConnectionBundle
-from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import RequestResponseDelay, SdClientConfig
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.ObsoleteModel import SocketConnection, SocketConnectionBundle
+from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetTopology import ApplicationEndpoint, RequestResponseDelay, SdClientConfig
 
 
 class TransportProtocolConfiguration(ARObject, ABC):
@@ -1829,142 +1829,6 @@ class ProvidedServiceInstance(AbstractServiceInstance):
         """
         if value is not None:
             self.serviceIdentifier = value
-        return self
-
-
-class ApplicationEndpoint(Identifiable):
-    """An application endpoint is the endpoint on an Ecu in terms of application addressing (e.g. socket). The application endpoint represents e.g. the listen socket in client-server-based communication."""
-
-    # ApplicationEndpoint method parity checklist:
-    # Spec: AUTOSAR_CP_TPS_SystemTemplate.pdf, Table 6.124, p.458
-    # Columns: impl / docstring / test / reader / writer   ([—] = no XML element)
-    # [x] __init__                             [x] impl  [x] docstring  [x] test  [—] reader  [—] writer
-    # [x] createConsumedServiceInstance        [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] getConsumedServiceInstances          [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] getMaxNumberOfConnections            [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setMaxNumberOfConnections            [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] getNetworkEndpointRef                [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setNetworkEndpointRef                [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] getPriority                          [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setPriority                          [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] createProvidedServiceInstance        [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] getProvidedServiceInstances          [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] getTlsCryptoMappingRef               [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setTlsCryptoMappingRef               [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-    # [x] getTpConfiguration                   [x] impl  [x] docstring  [x] test  [—] reader  [x] writer
-    # [x] setTpConfiguration                   [x] impl  [x] docstring  [x] test  [x] reader  [—] writer
-
-    def __init__(self, parent: ARObject, short_name: str):
-        super().__init__(parent, short_name)
-
-        # Consumed service instances.
-        self.consumedServiceInstances: List[ConsumedServiceInstance] = []
-
-        # This attribute defines the maximal number of clients the Server is able to deal with in case of Service Discovery.
-        self.maxNumberOfConnections: Optional[PositiveInteger] = None
-
-        # Reference to the network address.
-        self.networkEndpointRef: Optional[RefType] = None
-
-        # Defines the frame priority where values from 0 (best effort) to 7 (highest) are allowed.
-        self.priority: Optional[PositiveInteger] = None
-
-        # Provided service instances.
-        self.providedServiceInstances: List[ProvidedServiceInstance] = []
-
-        # This reference identifies the applicable TlsCryptoServiceMapping that adds the ability for TLS-based encryption on the enclosing ApplicationEndpoint.
-        self.tlsCryptoMappingRef: Optional[RefType] = None
-
-        # Configuration of the used transport protocol.
-        self.tpConfiguration: Optional[TransportProtocolConfiguration] = None
-
-    def createConsumedServiceInstance(self, short_name: str) -> ConsumedServiceInstance:
-        """Consumed service instances."""
-        if not self.IsElementExists(short_name, ConsumedServiceInstance):
-            instance = ConsumedServiceInstance(self, short_name)
-            self.addElement(instance)
-            self.consumedServiceInstances.append(instance)
-        return self.getElement(short_name, ConsumedServiceInstance)
-
-    def getConsumedServiceInstances(self) -> List[ConsumedServiceInstance]:
-        """Consumed service instances."""
-        return self.consumedServiceInstances
-
-    def getMaxNumberOfConnections(self) -> Optional[PositiveInteger]:
-        """This attribute defines the maximal number of clients the Server is able to deal with in case of Service Discovery."""
-        return self.maxNumberOfConnections
-
-    def setMaxNumberOfConnections(self, value: Optional[PositiveInteger]) -> "ApplicationEndpoint":
-        """
-        This attribute defines the maximal number of clients the Server is able to deal with in case of Service Discovery.
-        A None value is a no-op and does not overwrite an existing maxNumberOfConnections.
-        """
-        if value is not None:
-            self.maxNumberOfConnections = value
-        return self
-
-    def getNetworkEndpointRef(self) -> Optional[RefType]:
-        """Reference to the network address."""
-        return self.networkEndpointRef
-
-    def setNetworkEndpointRef(self, value: Optional[RefType]) -> "ApplicationEndpoint":
-        """
-        Reference to the network address.
-        A None value is a no-op and does not overwrite an existing networkEndpointRef.
-        """
-        if value is not None:
-            self.networkEndpointRef = value
-        return self
-
-    def getPriority(self) -> Optional[PositiveInteger]:
-        """Defines the frame priority where values from 0 (best effort) to 7 (highest) are allowed."""
-        return self.priority
-
-    def setPriority(self, value: Optional[PositiveInteger]) -> "ApplicationEndpoint":
-        """
-        Defines the frame priority where values from 0 (best effort) to 7 (highest) are allowed.
-        A None value is a no-op and does not overwrite an existing priority.
-        """
-        if value is not None:
-            self.priority = value
-        return self
-
-    def createProvidedServiceInstance(self, short_name: str) -> ProvidedServiceInstance:
-        """Provided service instances."""
-        if not self.IsElementExists(short_name, ProvidedServiceInstance):
-            instance = ProvidedServiceInstance(self, short_name)
-            self.addElement(instance)
-            self.providedServiceInstances.append(instance)
-        return self.getElement(short_name, ProvidedServiceInstance)
-
-    def getProvidedServiceInstances(self) -> List[ProvidedServiceInstance]:
-        """Provided service instances."""
-        return self.providedServiceInstances
-
-    def getTlsCryptoMappingRef(self) -> Optional[RefType]:
-        """This reference identifies the applicable TlsCryptoServiceMapping that adds the ability for TLS-based encryption on the enclosing ApplicationEndpoint."""
-        return self.tlsCryptoMappingRef
-
-    def setTlsCryptoMappingRef(self, value: Optional[RefType]) -> "ApplicationEndpoint":
-        """
-        This reference identifies the applicable TlsCryptoServiceMapping that adds the ability for TLS-based encryption on the enclosing ApplicationEndpoint.
-        A None value is a no-op and does not overwrite an existing tlsCryptoMappingRef.
-        """
-        if value is not None:
-            self.tlsCryptoMappingRef = value
-        return self
-
-    def getTpConfiguration(self) -> Optional[TransportProtocolConfiguration]:
-        """Configuration of the used transport protocol."""
-        return self.tpConfiguration
-
-    def setTpConfiguration(self, value: Optional[TransportProtocolConfiguration]) -> "ApplicationEndpoint":
-        """
-        Configuration of the used transport protocol.
-        A None value is a no-op and does not overwrite an existing tpConfiguration.
-        """
-        if value is not None:
-            self.tpConfiguration = value
         return self
