+++ tests/test_armodel/models/M2/AUTOSARTemplates/SystemTemplate/Fibex/Fibex4Ethernet/test_CanXlProps.py
@@ -10,12 +10,10 @@ import pytest
 
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import PositiveInteger
 from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Can.CanTopology import (
+    CanControllerConfiguration,
     CanControllerFdConfiguration,
     CanControllerXlConfiguration,
     CanControllerXlConfigurationRequirements,
-)
-from armodel.models.M2.AUTOSARTemplates.SystemTemplate.Fibex.Fibex4Ethernet.EthernetCommunication import (
-    CanControllerConfiguration,
     CanXlProps,
 )
